int to double

+++ src/FindRange.java
@@ -5,12 +5,12 @@ import java.util.List;
 
 public class FindRange {
 
-    private int x;
-    private int y;
-    HashMap<Integer , Integer> point = new HashMap<>();
-    ArrayList<Integer> yAxis = new ArrayList<>();
+    private double x;
+    private double y;
+    HashMap<Double , Double> point = new HashMap<>();
+    ArrayList<Double> yAxis = new ArrayList<>();
 
-    public FindRange(int y , int x){
+    public FindRange(double y , double x){
         this . x = x;
         this. y = y;
     }
@@ -18,14 +18,14 @@ public class FindRange {
 
     }
 
-    public void put(int y , int x){
+    public void put(double y , double x){
         point.put(y , x);
         yAxis.add(y);
     }
 
-    public ArrayList<Integer> range(int minY , int maxY){
+    public ArrayList<Double> range(double minY , double maxY){
 
-        ArrayList<Integer> list = new ArrayList<>();
+        ArrayList<Double> list = new ArrayList<>();
 
         for (int i = 0 ; i < yAxis.size() ; i++){
             if (yAxis.get(i) <= maxY && yAxis.get(i) >= minY){
@@ -37,7 +37,7 @@ public class FindRange {
         return list;
     }
 
-    public int getX(int y){
+    public double getX(double y){
         return point.get(y);
     }
